[IDLE-61] refactor#82: 디렉토리 변경
- presentation, domain, persistence 구조로 개편
